Separando logica do DAO. Deixando o service referenciando uma interface, para que quando for realizar a instanciacao seja passada o tipo de banco desejado (sqlite, postgres)
